drop cargo for cargo_metadata
Cargo_metadata is a crate the expose the cargo-metadata subcommand. Rely
on metadata and not on cargo directly reduces considerably both the
runtime and compile-time costs.  The downside is that cargo_metadata
doesn't expose cargo's internal types and metadata. By now is not a
blocking issue because the missing informations can be extracted
directly from the manifest file.  Has been added also another argument
to the `build` to specify a manifest path, because of that now
cargo-ebuild doesn't rely anymore on a cargo project but only on the
manifest.  This change removes also some boilerplate code like resolve,
workspace, etc.

In future work should ported the [package.metadata] field of the
`Cargo.toml` to handle cargo-ebuild specific configurations (es
KEYWORDS).

In future new version of the cargo-metadata format and cargo features
should be mapped into cargo_metadata in order to be used.

Signed-off-by: gibix <gibix@riseup.net>