include license info for depends
Since dependencies are compiled into the final binary, those licenses
must be included in the ebuild generated by the tool. Gather up the
licenses from each of the dependencies and include them in an AND
pattern since computing the OR pattern as well is a bit more effort but
this is an improvement.