feat: include all features when pulling metadata
Optional package dependencies are not included when pulling metadata. In
some cases these packages are still required to actually build even when
those flags are turned off.

This change effectively adds `--all-features` to the call of cargo
metadata.

I don't know if this is due to a change in behavior from cargo build,
however, it seems prudent to include all possible dependencies.